Move scheduler and worker out of lib
